Add bootcamp ownership check

+++ controllers/bootcamps.js
@@ -34,10 +34,10 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 
 
     // Check for published bootcamp
-    const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id })
+    const publishedBootcamp = await Bootcamp.findOne({user: req.user.id})
 
     // Check user role, publisher can add only one
-    if (publishedBootcamp && req.user.role !== "admin"){
+    if (publishedBootcamp && req.user.role !== "admin") {
         return next(new ErrorResponse(`The user with ID ${req.user.id} has already published a bootcamp`, 400))
     }
 
@@ -50,18 +50,22 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route   UPDATE /api/v1/bootcamps/:id
 // @access  Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const updatedBootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    })
+    let bootcamp = await Bootcamp.findById(req.params.id)
 
-    if (!updatedBootcamp) {
+    if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
     }
 
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User with id of ${req.params.id} is not authorized to update this bootcamp`, 401));
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {new: true, runValidators: true})
+
     res.status(200).json({
         success: true,
-        data: updatedBootcamp
+        data: bootcamp
     });
 });
 
@@ -73,6 +77,11 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User with id of ${req.params.id} is not authorized to delete this bootcamp`, 401));
     }
 
     await bootcamp.deleteOne();
@@ -111,6 +120,11 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
     }
 
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User with id of ${req.params.id} is not authorized to upload the photo of this bootcamp`, 401));
+    }
+
     if (!req.files) {
         return next(new ErrorResponse("Please upload a file", 400))
     }
@@ -138,7 +152,7 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
         await Bootcamp.findByIdAndUpdate(req.params.id, {photo: file.name});
 
-        res.status(200).json({success:true, data:file.name})
+        res.status(200).json({success: true, data: file.name})
     })
 })
